+++ torch_geometric/nn/models/__init__.py
@@ -21,6 +21,7 @@ from .rect import RECT_L
 from .linkx import LINKX
 from .lightgcn import LightGCN
 from .mask_label import MaskLabel
+from .rev_gnn import GroupAddRev
 
 __all__ = [
     'MLP',
@@ -55,6 +56,7 @@ __all__ = [
     'LINKX',
     'LightGCN',
     'MaskLabel',
+    'GroupAddRev',
 ]
 
 classes = __all__